# Add support for writing a JsonObject and array-based overload