--- a/tests/Benchmarks/System.Text.JsonLab/JsonParserChangeEntryPerf.cs
+++ b/tests/Benchmarks/System.Text.JsonLab/JsonParserChangeEntryPerf.cs
@@ -5,9 +5,11 @@ using BenchmarkDotNet.Attributes;
 using Benchmarks;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using System.Buffers.Text;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Text.Formatting;
 
 namespace System.Text.JsonLab.Benchmarks
 {
@@ -26,6 +28,12 @@ namespace System.Text.JsonLab.Benchmarks
         private MemoryStream _stream;
         private StreamReader _reader;
 
+        private byte[] _output;
+        private MemoryStream _outputStream;
+        private StreamWriter _writer;
+
+        private ArrayFormatterWrapper _arrayOutput;
+
         [ParamsSource(nameof(TestCaseValues))]
         public TestCaseType TestCase;
 
@@ -39,6 +47,12 @@ namespace System.Text.JsonLab.Benchmarks
 
             _stream = new MemoryStream(_dataUtf8);
             _reader = new StreamReader(_stream, Encoding.UTF8, false, 1024, true);
+
+            _output = new byte[_dataUtf8.Length];
+            _outputStream = new MemoryStream(_output);
+            _writer = new StreamWriter(_outputStream, Encoding.UTF8, 1024, true);
+
+            _arrayOutput = new ArrayFormatterWrapper(1024, SymbolTable.InvariantUtf8);
         }
 
         [Benchmark(Baseline = true)]
@@ -47,7 +61,7 @@ namespace System.Text.JsonLab.Benchmarks
             _stream.Seek(0, SeekOrigin.Begin);
             TextReader reader = _reader;
             JToken deps;
-            using (JsonTextReader jsonReader = new JsonTextReader(reader))
+            using (var jsonReader = new JsonTextReader(reader))
             {
                 deps = JToken.ReadFrom(jsonReader);
             }
@@ -64,6 +78,13 @@ namespace System.Text.JsonLab.Benchmarks
 
             JProperty library = deps["libraries"].Children<JProperty>().First();
             library.Remove();
+
+            _outputStream.Seek(0, SeekOrigin.Begin);
+            TextWriter writer = _writer;
+            using (var jsonWriter = new JsonTextWriter(writer))
+            {
+                deps.WriteTo(jsonWriter);
+            }
         }
 
         [Benchmark]
@@ -81,6 +102,13 @@ namespace System.Text.JsonLab.Benchmarks
             JsonObject libraries = obj["libraries"];
             if (libraries.TryGetChild(out child))
                 obj.Remove(child);
+
+            _arrayOutput.Clear();
+            var jsonUtf8 = new Utf8JsonWriter<ArrayFormatterWrapper>(_arrayOutput);
+            jsonUtf8.Write(obj);
+            jsonUtf8.Flush();
+            
+            obj.Dispose();
         }
     }
 }