--- a/tests/Benchmarks/System.Text.JsonLab/JsonWriterPerf.cs
+++ b/tests/Benchmarks/System.Text.JsonLab/JsonWriterPerf.cs
@@ -2,20 +2,21 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using BenchmarkDotNet.Attributes;
-using BenchmarkDotNet.Diagnostics.Windows.Configs;
 using System.Buffers.Text;
 using System.IO;
 using System.Text.Formatting;
 
 namespace System.Text.JsonLab.Benchmarks
 {
-    [SimpleJob(-1, 5, 10, 32768)]
-    [DisassemblyDiagnoser(printAsm: true, printSource: true)]
-    [InliningDiagnoser()]
+    [SimpleJob(-1, 3, 5)]
     [MemoryDiagnoser]
     public class JsonWriterPerf
     {
-        private const int ExtraArraySize = 1000;
+        private static readonly byte[] Message = Encoding.UTF8.GetBytes("message");
+        private static readonly byte[] HelloWorld = Encoding.UTF8.GetBytes("Hello, World!");
+        private static readonly byte[] ExtraArray = Encoding.UTF8.GetBytes("ExtraArray");
+
+        private const int ExtraArraySize = 10000000;
         private const int BufferSize = 1024 + (ExtraArraySize * 64);
 
         private ArrayFormatterWrapper _arrayFormatterWrapper;
@@ -94,6 +95,10 @@ namespace System.Text.JsonLab.Benchmarks
         [Arguments(10)]
         [Arguments(100)]
         [Arguments(1000)]
+        [Arguments(10000)]
+        [Arguments(100000)]
+        [Arguments(1000000)]
+        [Arguments(10000000)]
         public void WriterSystemTextJsonArrayOnly(int size)
         {
             _arrayFormatterWrapper.Clear();
@@ -107,6 +112,10 @@ namespace System.Text.JsonLab.Benchmarks
         [Arguments(10)]
         [Arguments(100)]
         [Arguments(1000)]
+        [Arguments(10000)]
+        [Arguments(100000)]
+        [Arguments(1000000)]
+        [Arguments(10000000)]
         public void WriterUtf8JsonArrayOnly(int size)
         {
             WriterUtf8JsonArrayOnly(_data.AsSpan(0, size), _output);
@@ -135,14 +144,7 @@ namespace System.Text.JsonLab.Benchmarks
             json.WriteAttribute("city", "Redmond");
             json.WriteAttribute("zip", 98052);
             json.WriteObjectEnd();
-
-            json.WriteArrayStart("ExtraArray");
-            for (var i = 0; i < data.Length; i++)
-            {
-                json.WriteValue(data[i]);
-            }
-            json.WriteArrayEnd();
-
+            json.WriteArrayUtf8(ExtraArray, data);
             json.WriteObjectEnd();
             json.Flush();
         }
@@ -240,7 +242,7 @@ namespace System.Text.JsonLab.Benchmarks
             Utf8JsonWriter<ArrayFormatterWrapper> json = Utf8JsonWriter.Create(output, formatted);
 
             json.WriteObjectStart();
-            json.WriteAttribute("message", "Hello, World!");
+            json.WriteAttributeUtf8(Message, HelloWorld);
             json.WriteObjectEnd();
             json.Flush();
         }
@@ -272,12 +274,7 @@ namespace System.Text.JsonLab.Benchmarks
         {
             Utf8JsonWriter<ArrayFormatterWrapper> json = Utf8JsonWriter.Create(output, formatted);
 
-            json.WriteArrayStart("ExtraArray");
-            for (var i = 0; i < data.Length; i++)
-            {
-                json.WriteValue(data[i]);
-            }
-            json.WriteArrayEnd();
+            json.WriteArrayUtf8(ExtraArray, data);
             json.Flush();
         }